fix warning

+++ include/fluid/fluid.h
@@ -92,7 +92,7 @@ struct FluidConfiguration {
     /**
      * @brief max angle ardupilot parameter for the travel operation.
      */
-    const float travel_max_angle;  
+    const int travel_max_angle;  
 
     /**
      * @brief 3D offset of the Face_hugger compared to the drone center
